+++ src/blob/index/simple.rs
@@ -9,7 +9,10 @@ pub(crate) struct SimpleFileIndex {
 }
 
 #[async_trait::async_trait]
-impl<K: Key> FileIndexTrait<K> for SimpleFileIndex {
+impl<K> FileIndexTrait<K> for SimpleFileIndex
+where
+    for<'a> K: Key<'a>,
+{
     async fn from_file(name: FileName, ioring: Option<Rio>) -> Result<Self> {
         trace!("open index file");
         let file = File::open(name.to_path(), ioring)
@@ -144,11 +147,14 @@ impl SimpleFileIndex {
         IndexHeader::from_raw(&buf).map_err(Into::into)
     }
 
-    async fn search_all<K: Key>(
+    async fn search_all<K>(
         file: &File,
         key: &K,
         index_header: &IndexHeader,
-    ) -> Result<Option<Vec<RecordHeader>>> {
+    ) -> Result<Option<Vec<RecordHeader>>>
+    where
+        for<'a> K: Key<'a>,
+    {
         if let Some(header_pos) = Self::binary_search(file, key, index_header)
             .await
             .with_context(|| "blob, index simple, search all, binary search failed")?
@@ -217,11 +223,14 @@ impl SimpleFileIndex {
         }
     }
 
-    async fn binary_search<K: Key>(
+    async fn binary_search<K>(
         file: &File,
         key: &K,
         header: &IndexHeader,
-    ) -> Result<Option<(RecordHeader, usize)>> {
+    ) -> Result<Option<(RecordHeader, usize)>>
+    where
+        for<'a> K: Key<'a>,
+    {
         debug!("blob index simple binary search header {:?}", header);
 
         let mut start = 0;
@@ -235,7 +244,7 @@ impl SimpleFileIndex {
                 "blob index simple binary search mid header: {:?}",
                 mid_record_header
             );
-            let cmp = key.cmp(&mid_record_header.key().to_vec().into());
+            let cmp = key.as_ref_key().cmp(&mid_record_header.key().into());
             debug!("mid read: {:?}, key: {:?}", mid_record_header.key(), key);
             debug!("before mid: {:?}, start: {:?}, end: {:?}", mid, start, end);
             match cmp {
@@ -255,7 +264,10 @@ impl SimpleFileIndex {
         Ok(None)
     }
 
-    async fn validate_header<K: Key>(&self, buf: &mut Vec<u8>) -> Result<()> {
+    async fn validate_header<K>(&self, buf: &mut Vec<u8>) -> Result<()>
+    where
+        for<'a> K: Key<'a>,
+    {
         FileIndexTrait::<K>::validate(self)?;
         if !Self::hash_valid(&self.header, buf)? {
             let param = ValidationErrorKind::IndexChecksum;
@@ -264,10 +276,13 @@ impl SimpleFileIndex {
         Ok(())
     }
 
-    fn serialize<K: Key>(
+    fn serialize<K>(
         headers: &InMemoryIndex<K>,
         meta: Vec<u8>,
-    ) -> Result<Option<(IndexHeader, Vec<u8>)>> {
+    ) -> Result<Option<(IndexHeader, Vec<u8>)>>
+    where
+        for<'a> K: Key<'a>,
+    {
         debug!("blob index simple serialize headers");
         if let Some(record_header) = headers.values().next().and_then(|v| v.first()) {
             debug!("index simple serialize headers first: {:?}", record_header);