+++ src/filter/mod.rs
@@ -1,5 +1,6 @@
 /// Bloom filter
 pub mod bloom;
+mod combined;
 /// Hierarchical
 pub mod hierarchical;
 /// Range filter
@@ -9,8 +10,8 @@ pub mod traits;
 
 use super::prelude::*;
 pub use bloom::*;
+pub use combined::*;
 pub use hierarchical::*;
-pub use range::*;
 use std::ops::Add;
 pub use traits::*;